rag-doc-qa/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
openai==1.43.0
python-dotenv==1.0.1